% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_groups_by.R
\name{balance_groups_by}
\alias{balance_groups_by}
\alias{balance_groups_by.default}
\alias{balance_groups_by.goldrake}
\title{Set balance classes}
\usage{
balance_groups_by(x, ...)

\method{balance_groups_by}{default}(x, ...)

\method{balance_groups_by}{goldrake}(x, ..., force = FALSE)
}
\arguments{
\item{x}{a \code{\link{goldrake}} object}

\item{...}{variables to balance by}

\item{force}{(lgl, default = TRUE) if TRUE, overwrite possible
balancing variables already set.}
}
\value{
invisibly, the updated version of x
}
\description{
Set balance classes
}
\examples{
goldrake(mtcars) \%>\%
    balance_groups_by(vs)

goldrake(mtcars) \%>\%
    balance_groups_by(vs, cyl)

goldrake(mtcars) \%>\%
    balance_groups_by(vs, cyl)

goldrake(mtcars) \%>\%
    balance_groups_by(vs) \%>\%
    balance_groups_by(vs, cyl, force = TRUE)

\dontrun{
    # an error
    goldrake(mtcars) \%>\%
        balance_groups_by(foo)

   goldrake(mtcars) \%>\%
        balance_groups_by(vs) \%>\%
        balance_groups_by(vs, cyl)
}

}
