% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_setup.R
\name{confirm_setup}
\alias{confirm_setup}
\alias{confirm_setup.goldrake}
\title{Confirm goldrake setup}
\usage{
confirm_setup(x, prj_name = "goldrake")

\method{confirm_setup}{goldrake}(x, prj_name = "goldrake")
}
\arguments{
\item{x}{a goldrake object}

\item{prj_name}{(chr, default = "goldrake") name for the project}
}
\value{
invisibly(x), moreover it save a local copy of goldrake
}
\description{
Confirm goldrake setup
}
\examples{
\dontrun{
    library(goldrake)
    gold <- goldrake(mtcars) \%>\%
        set_gold_classes() \%>\%
        add_reviewer("Rev 1")

    confirm_setup(gold)
}
}
