% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldrake-package.R
\docType{package}
\name{goldrake-package}
\alias{goldrake}
\alias{goldrake-package}
\title{goldrake: Gold-standard hero}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
The goal of goldrake is to provide an environment
    to create gold-standard databases for a classification task.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CorradoLanera/goldrake}
  \item Report bugs at \url{https://github.com/CorradoLanera/goldrake/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com}

Authors:
\itemize{
  \item Daniele Bottigliengo \email{daniele.bottigliengo@libero.it}
}

}
\keyword{internal}
