% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldrake.R, R/print-methods.R
\name{goldrake}
\alias{goldrake}
\alias{print.goldrake}
\title{Goldrakes}
\usage{
goldrake(original_data)

\method{print}{goldrake}(x, ...)
}
\arguments{
\item{original_data}{a data frame like object with the original
dataset.}

\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\value{
a \link[goldrake:goldrake-package]{goldrake} object.

invisible \code{x}
}
\description{
The function \code{\link{goldrake}} creates goldrake objects:
collection of data, classes defined by expert reviewers, and
sampling characteristics used to create a gold standard data set
for classification tasks.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Method to print \code{\link{goldrake}}s
}}

\examples{
goldrake(mtcars)
\dontrun{
    # throw an error
    goldrake(1)
}
x <- goldrake(mtcars)
}
