% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldrake_list.R
\name{goldrake_list}
\alias{goldrake_list}
\title{goldrake list component}
\usage{
goldrake_list(.x)
}
\arguments{
\item{.x}{a data.frame.}
}
\value{
a list with three components: \code{used_data} (an empty
\link[tibble:tibble-package]{tibble}), \code{reviewers} (an empty
list), and \code{original_data} (.x as the original data).
}
\description{
This function create the list-component of a goldrake object
}
