% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print}
\alias{print}
\title{Print goldrakes}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\value{
invisible(x)
}
\description{
\code{print} prints its argument and returns it invisibly
(via invisible(x))
}
