% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gold_class.R
\name{set_gold_classes}
\alias{set_gold_classes}
\alias{set_gold_classes.default}
\alias{set_gold_classes.goldrake}
\title{Set gold classes}
\usage{
set_gold_classes(x, classes, ...)

\method{set_gold_classes}{default}(x, classes, ...)

\method{set_gold_classes}{goldrake}(x, classes = c("0", "1"), ...,
  force = FALSE)
}
\arguments{
\item{x}{a \code{\link{goldrake}} object}

\item{classes}{(chr, default c("0", "1")) Classes admitted for the
classification.}

\item{...}{further arguments passed to the methods}

\item{force}{(lgl, default = TRUE) if TRUE, overwrite possible
classes already set.}
}
\value{
invisibly, a \code{\link{goldrake}} object with updated
classes
}
\description{
Set gold classes
}
\examples{
goldrake(mtcars) \%>\%
    set_gold_classes()

goldrake(mtcars) \%>\%
    set_gold_classes(c("good", "bad", "so and so"))

goldrake(mtcars) \%>\%
    set_gold_classes(c("good", "bad", "so and so")) \%>\%
    set_gold_classes(c("0", "1"))

goldrake(mtcars) \%>\%
    set_gold_classes(c("good", "bad", "so and so")) \%>\%
    set_gold_classes(c("0", "1"), force = TRUE)

\dontrun{
    # an error
    set_gold_classes(1)

    # another error
    goldrake(mtcars) \%>\%
        set_gold_classes(1)
}

}
