% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_classify.R
\name{start_classify}
\alias{start_classify}
\alias{start_classify.NULL}
\alias{start_classify.goldrake}
\title{Start to classify}
\usage{
start_classify(x = NULL, gold_dir = NULL, gold_name = NULL)

\method{start_classify}{NULL}(x = NULL, gold_dir = NULL,
  gold_name = NULL)

\method{start_classify}{goldrake}(x, gold_dir = NULL, gold_name = NULL)
}
\arguments{
\item{x}{a \code{\link{goldrake}}}

\item{gold_dir}{(chr, default = NULL) path to the project directory.
If not provided the function automatically use (or create, if
it not exists) the folder \code{.goldrake} in the current
working directory.}

\item{gold_name}{(chr, default = NULL) the name of the project.}
}
\value{
the update version of \code{\link{goldrake}} x
}
\description{
start to classify a \code{\link{goldrake}}.
}
\examples{
\dontrun{
    gold <- goldrake(mtcars) \%>\%
        set_gold_classes() \%>\%
        add_reviewer("Rev 1") \%>\%
        confirm_setup()

    start_classify(gold)
}
}
