% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_reviewer.R
\name{add_reviewer}
\alias{add_reviewer}
\alias{add_reviewer.default}
\alias{add_reviewer.goldrake}
\title{Add reviewer}
\usage{
add_reviewer(x, name, surname = NULL, inspector = FALSE, seed = 1)

\method{add_reviewer}{default}(x, name, surname = NULL,
  inspector = FALSE, seed = 1)

\method{add_reviewer}{goldrake}(x, name, surname = NULL,
  inspector = FALSE, seed = 1)
}
\arguments{
\item{x}{a \code{\link{goldrake}} object}

\item{name}{(chr) Reviewer's name}

\item{surname}{(chr, default = NULL, optional) Reviewer's surname}

\item{inspector}{(lgl, default = FALSE) Is a reviewer a reviever's
reviewers?}

\item{seed}{(num, default = 1) seed for the case sampling}
}
\value{
invisibly, the updated version of x
}
\description{
Add reviewer
}
\examples{
with_reviewer <- goldrake(mtcars) \%>\%
    set_gold_classes() \%>\%
    add_reviewer("Reviewer 1")

with_reviewer
}
