% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldrake.R
\name{goldrake}
\alias{goldrake}
\title{Goldrakes}
\usage{
goldrake(original_data)
}
\arguments{
\item{original_data}{a data frame like object with the original
dataset.}
}
\value{
a \link[goldrake:goldrake-package]{goldrake} object.
}
\description{
The function \code{\link{goldrake}} creates goldrake objects:
collection of data, classes defined by expert reviewers, and
sampling characteristics used to create a gold standard data set
for classification tasks.
}
\examples{
goldrake(mtcars)
}
